move generic api types to separate file
