+++ src/services/api/genericApi.helpers.ts
@@ -1,15 +1,8 @@
-export class BaseHeaders {
-  public 'Accept': string;
-  public 'Content-Type': string;
-  public 'observe': string;
-}
-
-export type TokenOptions = 'token' | 'no-token' | 'only-token' ;
-export enum TokenOptionsEnum {
-  token = 'token',
-  noToken = 'no-token',
-  onlyToken = 'only-token',
-}
+import {
+  BaseHeaders,
+  TokenOptions,
+  TokenOptionsEnum
+} from './genericApi.types';
 
 export const baseHeaders: BaseHeaders =  {
   'Accept': 'application/json',
